pyproject.toml
[tool.poetry]
name = "toy-blog"
version = "0.1.0"
description = "A simple Blog App with Django"
authors = ["Stefan Cosic <stefan22bl@gmail.com>"]
readme = "README.md"

[tool.flake8]
exclude = "settings.py"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.2"
djangorestframework = "^3.15.2"
black = {extras = ["d"], version = "^24.10.0"}
psycopg2 = "^2.9.10"
django-environ = "^0.11.2"
flake8 = "^7.1.1"
nox = "^2024.10.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
